+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Inspector/ToonLitConvertSettingsDrawer.cs
@@ -19,14 +19,23 @@ namespace KRT.VRCQuestTools.Inspector
         /// <inheritdoc />
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
-            var i18n = VRCQuestToolsSettings.I18nResource;
-
             if (label.text != string.Empty || label.image != null)
             {
                 EditorGUILayout.LabelField(label);
             }
 
             EditorGUI.indentLevel++;
+            DrawPorpertyFields(property);
+            EditorGUI.indentLevel--;
+        }
+
+        /// <summary>
+        /// Draw property fields for ToonLitConvertSettings.
+        /// </summary>
+        /// <param name="property">Serialized property of ToonLitConvertSettings.</param>
+        internal static void DrawPorpertyFields(SerializedProperty property)
+        {
+            var i18n = VRCQuestToolsSettings.I18nResource;
 
             var generateQuestTextures = property.FindPropertyRelative("generateQuestTextures");
             EditorGUILayout.PropertyField(generateQuestTextures, new GUIContent(i18n.GenerateQuestTexturesLabel));
@@ -34,9 +43,8 @@ namespace KRT.VRCQuestTools.Inspector
             {
                 EditorGUILayout.PropertyField(property.FindPropertyRelative("maxTextureSize"), new GUIContent(i18n.IMaterialConvertSettingsTexturesSizeLimitLabel));
                 EditorGUILayout.PropertyField(property.FindPropertyRelative("mainTextureBrightness"), new GUIContent(i18n.IMaterialConvertSettingsMainTextureBrightnessLabel, i18n.IMaterialConvertSettingsMainTextureBrightnessTooltip));
+                EditorGUILayout.PropertyField(property.FindPropertyRelative("generateShadowFromNormalMap"), new GUIContent(i18n.ToonLitConvertSettingsGenerateShadowFromNormalMapLabel));
             }
-
-            EditorGUI.indentLevel--;
         }
     }
 }